kietOLX-backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.12.0